requirements.txt
line-bot-sdk>=3.0.0,<4.0.0
schedule>=1.2.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
